# Avoid lambda compilation for constant and member access expressions in Criteria