```C#
//Example of constant
query.Where(x => x.Id == 1);

//Example of member access
var paramId = 1;
query.Where(x => x.Id == paramId)
```