@hazzik 2020-07-24
@bahusoid is performance gain significant?