@bahusoid 2020-07-25
@hazzik Benchmark for slower member access case (`Where(x => x.Id == paramId)`):
```
| Method |       Mean |     Error |     StdDev | Ratio |
|------- |-----------:|----------:|-----------:|------:|
|    New |   1.562 us | 0.0283 us |  0.0236 us | 0.007 |
|    Old | 233.980 us | 4.6614 us | 12.1980 us | 1.000 |

  1 us   : 1 Microsecond (0.000001 sec)
```

Source: https://gist.github.com/bahusoid/c68cb6b81f8dde9252d0cc67229136b4